+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
     * Port Dry Dock field uses the proper enumeration for display
     * Delete Electronic Publications which are not returned from the API 
     * Electronic Publications now sync once a day instead of once a week
+    * Display an error upon failed download of an Electronic Publication
 
 ## 1.0.1
   ### Bug Fixes